requirements.txt
bert-extractive-summarizer==0.4.2
Flask==1.1.2
gunicorn==20.0.4
h5py==2.10.0
PyPDF2==1.26.0
torch==1.5.1
torchvision==0.6.1
